Setup first iteration of navigation panel
It should now be possible to return to login and signup from DataTypesUI via the navigation panel. So long as no major changes are made to the scene it should work just fine. Working on a script to replace the current functionality, to make it more reliable and less likely to break if the scene is changed or altered.