So that amounted to nothing. Great!
Navigation panel disappeared in pull immediately before pushing to origin. likely cause is renaming of scene elements, but whatever. Regardless, everything I just tried to push is gone.(Not a lot of wasted work, so who cares.) I'll just get to work on the script I mentioned in the previous push.